Recorder::stop is only called by Recorder::start, which seems wrong since whoever calls start would expect to call stop later.  Rename start() to do_it(), and include the stop() code at the end of do_it().
initHDF5 allocates something - a file - that never gets freed.  Define finishHDF5() to free
what initHDF5 allocates, and call it from main.

initHDF5 allocates a DataSpace that it only closes, and never frees, and does it with a local
pilot_dataspace variable that shadows the pilot_dataspace member of Recorder.  I think that
this local pilot_dataspace can be dropped, as long as the side-effect of its creation, a
call to file->createDataSet() is kept.


Former-commit-id: 5c0822d976734882f5a00dbe492e8d3e58ad1918